services/hypernode-gateway/requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
pyyaml==6.0.2
python-dotenv==1.0.1
httpx==0.27.2
itsdangerous==2.2.0

